Moving to event based phase changes
In the 'RoundController' you can set the 'Allow Timed Phases' which will adjust if you're using the event based or the time based phase changes.
Event based: If you are in the building phase you can trigger the combat phase with a UI button click. If you are in combat phase you should be changing to build phase if the last combatant dies.